Add files via upload
There are two parts to this project.

First you will implement the specified functions in fermat.py. This file focuses on the task of determining whether a number is prime or composite. (The Miller-Rabin test) and (Fermat Little Theorem) 

Second you will implement the specified functions in rsa.py. This file focuses on the task of generating a valid public-private keypair.